pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
exclude = [".idea", "tmp", "dist", ".tox", ".pytest_cache", "notebooks"]
name = "src"
version = "0.0.1.dev0"
authors = ["Marco Scattolin <scattolin.marco@bcg.com>"]
readme = "README.md"
description = ""


[tool.poetry.dependencies]
python = "3.11.*"
# ------------------ basic_vm ---------------------------------
python-dateutil = "2.8.2"
pytz = "2023.3"
PyYAML = "6.0"
six = "1.16.0"
typing_extensions = "4.6.1"
furo = "2023.5.20"
traitlets = "5.9.0"
pydantic = "2.6.3"
pydantic-settings = "2.2.1"
numpy = "2.0.0"
pandas = "2.2.2"
tzdata = "2023.3"
scikit-learn = "1.5.1"
statsmodels = "0.14.2"
notebook = "6.5.4"
plotly = "5.15.0"
seaborn = "0.12.2"
lightgbm = "3.3.5"
xlrd = "2.0.1"
openpyxl = "3.1.2"
pyarrow = "17.0.0"
SQLAlchemy = "1.4.51"
psycopg2-binary = "2.9.9"
plotnine = "0.13.6"
pip-audit = "2.7.3"

# ------------------ s3 ---------------------------------
[tool.poetry.group.s3]
optional = true

[tool.poetry.group.s3.dependencies]
fsspec = "^2024.2.0"
s3fs = "^2024.2.0"


# ------------------ spark ----------------------------------
[tool.poetry.group.spark]
optional = true

[tool.poetry.group.spark.dependencies]
pyspark = "3.5.0"

# ------------------ airflow -------------------------
[tool.poetry.group.airflow]
optional = true

[tool.poetry.group.airflow.dependencies]
apache-airflow = "2.8.2"
apache-airflow-providers-telegram = "4.3.1"

# ------------------ optimization ---------------------------
[tool.poetry.group.optimization]
optional = true

[tool.poetry.group.optimization.dependencies]
ortools = "9.6.2534"

# ------------------ others ---------------------------
[tool.poetry.group.others]
optional = true

[tool.poetry.group.others.dependencies]
cowsay = "5.0"
pyodbc = "4.0.39"
streamlit = "1.37.0"
azure-core = "1.27.1"
azure-storage-blob = "12.16.0"
prophet = "1.1.4"
xgboost = "1.7.6"
neuralforecast = "1.5.0"
mlforecast = "0.7.4"
statsforecast = "1.5.0"
# hierarchicalforecast = "0.4.1" # compatible versions not available yet
# ydata-profiling = "4.9.0"
# pycaret = "3.3.2" # compatible versions not available yet

# ------------------ bayesian -----------------------
[tool.poetry.group.bayesian]
optional = true

[tool.poetry.group.bayesian.dependencies]
pymc = "5.10.4"
arviz = "0.19.0"
nutpie = "0.9.2"
pytensor = "2.18.6"
mkl = "^2024.0.0"
numpyro = "0.15.2"
mkl-service = "2.4.1"

# ------------------ testing -----------------------
[tool.poetry.group.testing]

[tool.poetry.group.testing.dependencies]
pytest = "7.1.2"
pytest-cov = "3.0.0"
pytest-lazy-fixture = "0.6.3"
flake8 = "3.9.*"
black = "22.6.0"
pre-commit = "2.20.0"
isort = "5.9.3"
sql-formatter = "0.6"
tox = "3.25.1"

# ------------------ docs -----------------------
[tool.poetry.group.docs]

[tool.poetry.group.docs.dependencies]
Sphinx = "7.0.1"
sphinx-toolbox = "3.5.0"
sphinx-copybutton = "0.5.2"

# ------------------ local_development -----------------------
[tool.poetry.group.local_development]

[tool.poetry.group.local_development.dependencies]
jupyterlab = "3.5.3"
jupyter = "1.0.0"

# ------------------ black -----------------------
[tool.black]
line-length = 88
target_version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.mypy_cache
    | \.tox
    | \.venv
    | data
    | sphinx
    | dist
    | docs
    | notebooks
    | sphinx
    | venv
  )/
)
'''
